integrations/sie_llamaindex/pyproject.toml
[project]
name = "sie-llamaindex"
version = "0.1.7"
description = "SIE integration for LlamaIndex"
requires-python = ">=3.10"
dependencies = [
    "sie-sdk>=0.1.0",
    "llama-index-core>=0.11",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "chromadb>=0.5",
    "llama-index-vector-stores-chroma>=0.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/sie_llamaindex"]
